requirements.txt
arrow==1.2.2
asgiref==3.5.0
async-timeout==4.0.2
click==8.1.2
Deprecated==1.2.13
Django==4.0.4
django-redis==5.2.0
djangorestframework==3.13.1
joblib==1.1.0
nltk==3.7
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.2
regex==2022.3.15
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
wrapt==1.14.0
